Fix permute issue
git-svn-id: svn+ssh://svn.code.sf.net/p/x3270/x3270/trunk@1200 12e4f418-770b-42d6-82dd-bd6eefd7344d

+++ x3270/Common/unicode.c
@@ -771,7 +771,7 @@ ebcdic_to_multibyte_x(ebc_t ebc, unsigned char cs, char mb[],
 	outbuf = mb;
 	outbytesleft = mb_len;
 	nc = iconv(i_u2mb, &inbuf, &inbytesleft, &outbuf, &outbytesleft);
-	if (nc == (size_t)-1 || inbytesleft == nu8) {
+	if (nc == (size_t)-1 || inbytesleft == (unsigned)nu8) {
 		mb[0] = '?';
 		mb[1] = '\0';
 		return 2;
